// src/runtime/lib/api/loaderTypes.ts
export interface IListRunLoaderOptions<T, D = Record<string, any>> {
  mockItems?: T[];
  isMock?: boolean;
  params?: Record<string, any>;
  expire?: number;
  data?: D;
}
// ...
export interface IListLoaderOptions<T, O> {
  prefix?: string;
  url: string;
  primary?: string;
  mockItems?: T[];
  getURL?: (opts: IListRunLoaderOptions<T, O>) => string;
  getRequestOptions?: (opts: IListRunLoaderOptions<T, O>) => AxiosRequestConfig;
}